Modify `sample_launchers`.
